Added OutputFrame Parameter
Replaced the outputFilePath parameter for now as I am developing this
node as a runtime Localisation implementation. Mapping can be accomplished seperatley by stitching the frame +  point cloud. However, I will likely add a
"mapping" functionality that allows the output of the map to a file in
the future
(similar to the way the algorithm was verified in the publication)